+++ src/player/playlist.h
@@ -1,6 +1,7 @@
 #ifndef SRC_PLAYER_PLAYLIST_H
 #define SRC_PLAYER_PLAYLIST_H
 
+#include <sys/types.h>
 #include <stddef.h>
 
 #include "types/song.h"
@@ -9,8 +10,10 @@
 void playlist_append_song(const struct song *song);
 void playlist_clear(void);
 
-size_t playlist_get_songs(const struct song **songs);
-size_t playlist_get_current_song(const struct song **song);
+/* returns number of songs */
+int playlist_get_songs(const struct song **songs);
+/* returns index of current song, -1 if no currrent */
+int playlist_get_current_song(const struct song **song);
 
 #endif /* #ifndef SRC_PLAYER_PLAYLIST_H */
